Change methods in generateAdditiveMaps_BiasedBoxcar()
Different way of assigning the weights - it now tries to minimise overlap
in some sense to avoid modes totally overlapping one another. Modes are now
placed in an iterative fashion, with the largest N chosen until they cover
e.g. 80% of the cortex. These are placed in a random fashion, but not
overlapping each other. Then, we take the next N modes and place them, but
ignoring the already placed modes. In this way, big areas tend to be
separate, but by lowering the % coverage parameter can make things more
random.